炸了

+++ src/main/java/com/qf/controller/BooksController.java
@@ -4,9 +4,11 @@ package com.qf.controller;
 import com.qf.config.ImgesConfig;
 import com.qf.constan.StateCode;
 import com.qf.controller.base.Base;
+import com.qf.entity.BookStates;
 import com.qf.entity.Books;
 import com.qf.entity.BooksWithBLOBs;
 import com.qf.entity.vo.BooksVo;
+import com.qf.mapper.BookStatesMapper;
 import com.qf.mapper.SystemMapper;
 import com.qf.service.IBooksService;
 import com.qf.util.FileUtils;
@@ -20,7 +22,10 @@ import org.springframework.web.multipart.MultipartFile;
 import javax.servlet.http.HttpServletRequest;
 import java.io.File;
 import java.io.IOException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.List;
+import java.util.Random;
 import java.util.UUID;
 
 @RestController
@@ -30,6 +35,8 @@ public class BooksController extends Base {
     IBooksService booksService;
     @Autowired
     ImgesConfig imagePath;
+    @Autowired
+    BookStatesMapper bookStatesMapper;
 
     //通过 ISBN 查询对应的书种,已测试
     @GetMapping("/selectBooksByIsbn")
@@ -42,7 +49,7 @@ public class BooksController extends Base {
         }
     }
 
-    //多条件查询,已测试
+    //多条件查询,已测试有漏洞
     @GetMapping("/selectBooksByCondition")
     public State<Object> selectBooksByCondition(Books books,Integer currentPage, Integer pageSize){
         Page<Books> page = booksService.selectBooksByCondition(books, currentPage, pageSize);
@@ -62,7 +69,25 @@ public class BooksController extends Base {
     //添加书本
     @PostMapping("/insertBooks")
     public State<Object> insertBooks(@RequestBody BooksWithBLOBs booksWithBLOBs){
+        if (StringUtils.isEmpty(booksWithBLOBs.getIsbn())) {
+            return packaging(StateCode.SUCCESS,"插入失败",booksWithBLOBs);
+        }
+        //设置创建时间
+        SimpleDateFormat format = new SimpleDateFormat("yyyy年MM月dd日 HH时mm分ss秒");
+        Date date = new Date();
+        String dateString = format.format(date);
+        booksWithBLOBs.setCreateTime(dateString);
         int i = booksService.insertBooks(booksWithBLOBs);
+        //根据条形码,创建在馆信息
+        Random random = new Random();
+        int n = random.nextInt(899)+100;
+        BookStates bookStates = new BookStates();
+        bookStates.setBarCode(n + booksWithBLOBs.getIsbn());
+        bookStates.setBookId(booksWithBLOBs.getBookId());
+        bookStates.setState((byte) 0);
+        bookStates.setBorrowNumber(new Short("0"));
+        bookStates.setIsDelete(1);
+        bookStatesMapper.insert(bookStates);
         if (i != 0){
             return packaging(StateCode.SUCCESS,"插入成功",i);
         }else {
@@ -131,8 +156,8 @@ public class BooksController extends Base {
     }*/
 
     //通过修改 is_delete 为0进行删除,已测试
-    @PutMapping("/deleteBook")
-    public State<Object> deleteBookByBookId(Integer bookId){
+    @PostMapping("/deleteBook")
+    public State<Object> deleteBookByBookId(@RequestParam("bookId") Integer bookId){
         int i = booksService.deleteBookByBookId(bookId);
         if (i != 0){
             return  packaging(StateCode.SUCCESS,"删除成功",i);
@@ -142,8 +167,17 @@ public class BooksController extends Base {
     }
 
 
-    @PutMapping("/deleteBatch")
-    public State<Object> deleteBatch(int[] bookIds){
+    @PostMapping("/deleteBatch")
+    public State<Object> deleteBatch(int []bookIds){
+    /*public State<Object> deleteBatch(String bookIds){
+        if(StringUtils.isEmpty(bookIds)){
+            return packaging(StateCode.FAIL,"删除失败",null);
+        }
+        String[] split = bookIds.split(",");
+        int[] bookids = new int[split.length];
+        for (int i =0;i < split.length;i++){
+            bookids[i] = Integer.parseInt(split[i]);
+        }*/
         int i = booksService.deleteBatch(bookIds);
         if (i == bookIds.length){
             return  packaging(StateCode.SUCCESS,"删除成功",i);
